Fixes for download speed and progress bar inside steam
