+++ Cargo.toml
@@ -15,7 +15,7 @@ json = "0.11.14"
 lazy_static = "1.3.0"
 psutil = "1.6.0"
 regex = "1"
-reqwest = { version = "0.9.19", features = ["rustls-tls"], default-features = false }
+reqwest = { version = "0.10.6", features = ["rustls-tls", "blocking", "gzip"], default-features = false }
 users = "0.9"
 xdg = "^2.1"
 tar = "0.4"
